pyproject.toml
[project]
name = "pomodoro-poetry"
version = "0.1.0"
description = ""
authors = [
    {name = "west_9045",email = "globov.motya@mail.ru"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi (>=0.128.0,<0.129.0)",
    "uvicorn (>=0.40.0,<0.41.0)",
    "pydantic-settings (>=2.12.0,<3.0.0)",
    "sqlalchemy (>=2.0.46,<3.0.0)",
    "alembic (>=1.18.1,<2.0.0)",
    "psycopg2 (>=2.9.11,<3.0.0)",
    "redis (>=7.1.0,<8.0.0)",
    "python-jose[cryptography] (>=3.5.0,<4.0.0)",
    "requests (>=2.32.5,<3.0.0)",
    "httpx (>=0.28.1,<0.29.0)",
    "asyncpg (>=0.31.0,<0.32.0)",
    "greenlet (>=3.3.1,<4.0.0)",
    "poethepoet (>=0.41.0,<0.42.0)",
    "pytest (>=9.0.2,<10.0.0)",
    "factory-boy (>=3.3.3,<4.0.0)",
    "pytest-factoryboy (>=2.8.1,<3.0.0)",
    "pytest-asyncio (>=1.3.0,<2.0.0)",
]

[tool.poetry]
packages = [{include = "pomodoro_poetry", from = "src"}]


[tool.poetry.group.dev.dependencies]
freezegun = "^1.5.5"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poe]
executor = "poetry"

[tool.poe.tasks]
serve = "uvicorn app.main:app --reload"
test = "pytest"

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "session"
